style: updated manual input component
