+++ src/components/assets/transfer/styles/xcm-bridge.scss
@@ -165,8 +165,8 @@
   border: 1px solid $astar-blue-dark;
   box-sizing: border-box;
   filter: drop-shadow(0px 1px 2px rgba(0, 0, 0, 0.05));
-  border-radius: 32px;
-  padding: 1px 12px;
+  border-radius: 4px;
+  padding: 2px 10px;
   color: $astar-blue-dark;
   font-weight: 500;
   font-size: 11px;
